updated audit log
updatet message formatting on message update.
added a check to see if message content actually got changed, this would some times run even though the only change to the message was that discord updated it automatically to display a website embed.
changed the user added/left message to display the username instead of linking the discord user it self (as this would not work if the user left the guild).